fix: prevent race conditions in request tracing
The documentation of net/http/httptrace.ClientTrace clearly states the
following:

> Functions may be called concurrently from different goroutines and some
> may be called after the request has completed or failed.

The implementation neglegted this detail and assumed that after RoundTrip
returns we can simply reset the timings and only the new timings will be
recorded from that point on.

This commit adjusts the round-tripper to clone the request for each try
as this is the only way to set the context for a request which carries
the trace and its callbacks. For cloning the original context (without
the previous trace) is used and a new trace is then attached to the
request. It also removes the reset function as it is no longer used.